fix: current date fixed comments
