+++ src/entities/posts/ui/post-reader/ui/header/current-date-block/styled.ts
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled'
+
+export const StyledCurrentDateBlock = styled.div`
+  display: grid;
+  grid-template-columns: 11px auto;
+  grid-gap: 10px;
+  margin-bottom: 64px;
+`
