+++ src/entities/posts/ui/post-reader/ui/header/styled.ts
@@ -1,5 +1,4 @@
 import styled from '@emotion/styled'
-
 import {
   Image,
   ImageWrapper as SharedImageWrapper,
@@ -9,13 +8,6 @@ import {
 export const StyledHeader = styled.header`
   display: grid;
   grid-template-rows: auto auto auto;
-`
-
-export const StyledCurrentDateBlock = styled.div`
-  display: grid;
-  grid-template-columns: 11px auto;
-  grid-gap: 10px;
-  margin-bottom: 64px;
 `
 
 // Первая строка - картинка, текст, описание
